backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crowe-research-copilot"
version = "0.1.0"
description = "Natural Products Analysis Platform with Chemistry Intelligence"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Michael Crowe", email = "michael.crowe@crowelogic.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "rdkit>=2023.9.1",
    "networkx>=3.2.1",
    "polars>=0.20.0",
    "pyarrow>=14.0.0",
    "scipy>=1.11.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.3.0",
    "pint>=0.23",
    "seaborn>=0.13.0",
    "matplotlib>=3.8.0",
    "plotnine>=0.13.0",
    "matchms==0.27.0",
    "numba>=0.59.0,<0.60.0",
    "spectrum-utils>=0.4.0",
    "pyteomics>=4.6",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    # "mordred>=1.2.0", # Incompatible with networkx>=3
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pandas>=2.1.0",
    "xlsxwriter>=3.1.0",
    "openpyxl>=3.1.0",
    "chembl-webresource-client>=0.10.0",
    "pubchempy>=1.0.4",
    "biopython>=1.83",
    "joblib>=1.3.0",
    "tqdm>=4.66.0",
    "numpy<2.0.0",
]

[project.urls]
Homepage = "https://github.com/MichaelCrowe11/chemistry-data-facto"
Repository = "https://github.com/MichaelCrowe11/chemistry-data-facto"
"Bug Tracker" = "https://github.com/MichaelCrowe11/chemistry-data-facto/issues"

[project.optional-dependencies]
bayes = ["pymc>=5.10", "arviz>=0.16"]
dev = [
  "black>=24.3",
  "ruff>=0.5",
  "mypy>=1.8",
  "pre-commit>=3.7",
  "pytest-cov>=4.1",
  "pytest-asyncio>=0.23"
]
all = ["np-copilot[bayes,dev]"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "ANN", "C4", "SIM", "TCH"]
target-version = "py310"
ignore = ["ANN101", "ANN102"]

[tool.pytest.ini_options]
addopts = "-q --cov=src/crowe_copilot --cov-report=html --cov-report=term"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[project.scripts]
crowe-copilot = "crowe_copilot.cli:app"

[project.entry-points."console_scripts"]
crc = "crowe_copilot.cli:app"
